lesson3 Filtration
