+++ src/app/App.js
@@ -1,31 +1,10 @@
-import React, { useState } from "react";
+import React from "react";
 import Users from "./components/users";
-import api from "./api";
 
 function App() {
-    const [users, setUsers] = useState(api.users.fetchAll());
-    const handleDelete = (userId) => {
-        setUsers(users.filter((user) => user._id !== userId));
-    };
-    const handleToggleBookMark = (id) => {
-        setUsers(
-            users.map((user) => {
-                if (user._id === id) {
-                    return { ...user, bookmark: !user.bookmark };
-                }
-                return user;
-            })
-        );
-        console.log(id);
-    };
-
     return (
         <div>
-            <Users
-                handleDelete={handleDelete}
-                onToggleBookMark={handleToggleBookMark}
-                users={users}
-            />
+            <Users />
         </div>
     );
 }
