+++ src/app/api/fake.api/professions.api.js
@@ -6,7 +6,6 @@ export const professionsObject = {
     actor: { _id: "67rdca3eeb7f6fgeed471824", name: "Актер" },
     cook: { _id: "67rdca3eeb7f6fgeed471829", name: "Повар" }
 };
-
 export const professions = [
     { _id: "67rdca3eeb7f6fgeed471818", name: "Доктор" },
     { _id: "67rdca3eeb7f6fgeed471820", name: "Официант" },
@@ -15,13 +14,13 @@ export const professions = [
     { _id: "67rdca3eeb7f6fgeed471824", name: "Актер" },
     { _id: "67rdca3eeb7f6fgeed471829", name: "Повар" }
 ];
-
 const fetchAll = () =>
     new Promise((resolve) => {
         window.setTimeout(function () {
-            resolve(professionsObject);
+            resolve(professions);
         }, 2000);
     });
+
 export default {
     fetchAll
 };
