+++ src/app/api/fake.api/user.api.js
@@ -1,5 +1,4 @@
 import { professionsObject as professions } from "./professions.api";
-
 const qualities = {
     tedious: {
         _id: "67rdca3eeb7f6fgeed471198",
