+++ src/app/api/index.js
@@ -1,4 +1,4 @@
-import * as users from "./fake.api/user.api";
+import users from "./fake.api/user.api";
 import professions from "./fake.api/professions.api";
 const API = {
     users,
