+++ src/app/components/user.jsx
@@ -12,7 +12,7 @@ const User = ({
     rate,
     handleDelete,
     bookmark,
-    onToggleBookMark
+    handleToggleBookMark
 }) => {
     return (
         <>
@@ -29,7 +29,7 @@ const User = ({
                 <td>
                     <BookMark
                         status={bookmark}
-                        onClick={() => onToggleBookMark(_id)}
+                        onClick={() => handleToggleBookMark(_id)}
                     />
                 </td>
                 <td>
@@ -48,13 +48,13 @@ const User = ({
 User.propTypes = {
     _id: PropTypes.string.isRequired,
     name: PropTypes.string.isRequired,
-    qualities: PropTypes.array,
+    qualities: PropTypes.array.isRequired,
     profession: PropTypes.object,
     completedMeetings: PropTypes.number.isRequired,
     rate: PropTypes.number.isRequired,
-    handleDelete: PropTypes.func.isRequired,
+    handleDelete: PropTypes.func,
     bookmark: PropTypes.bool,
-    onToggleBookMark: PropTypes.func.isRequired
+    handleToggleBookMark: PropTypes.func
 };
 
 export default User;
