+++ src/app/components/users.jsx
@@ -6,31 +6,47 @@ import User from "./user";
 import api from "../api";
 import GroupList from "./groupList";
 import SeachStatus from "./seachStatus";
-const Users = ({ users: allUsers, ...rest }) => {
+const Users = ({ ...rest }) => {
     const [currentPage, setCurrentPage] = useState(1);
     const [professions, setProfessions] = useState();
+    const [users, setUsers] = useState([]);
     const [selectedProf, setSelectedProf] = useState();
 
     const pageSize = 2;
     useEffect(() => {
         api.professions.fetchAll().then((data) => setProfessions(data));
+        api.users.fetchAll().then((data) => setUsers(data));
     }, []);
     useEffect(() => {
         setCurrentPage(1);
     }, [selectedProf]);
 
+    const handleDelete = (userId) => {
+        setUsers(users.filter((user) => user._id !== userId));
+    };
+
+    const handleToggleBookMark = (id) => {
+        setUsers(
+            users.map((user) => {
+                if (user._id === id) {
+                    return { ...user, bookmark: !user.bookmark };
+                }
+                return user;
+            })
+        );
+        console.log(id);
+    };
+
     const handelProfessionSelect = (item) => {
         setSelectedProf(item);
     };
-
     const handlePageChange = (pageIndex) => {
         setCurrentPage(pageIndex);
         console.log("page: ", pageIndex);
     };
-
     const filteredUsers = selectedProf
-        ? allUsers.filter((user) => user.profession === selectedProf)
-        : allUsers;
+        ? users.filter((user) => user.profession._id === selectedProf._id)
+        : users;
     const count = filteredUsers.length;
     const userCrop = paginate(filteredUsers, currentPage, pageSize);
     const clearFilter = () => {
@@ -81,6 +97,8 @@ const Users = ({ users: allUsers, ...rest }) => {
                         pageSize={pageSize}
                         currentPage={currentPage}
                         onPageChange={handlePageChange}
+                        handleDelete={handleDelete}
+                        handleToggleBookMark={handleToggleBookMark}
                     />
                 </div>
             </div>
@@ -89,7 +107,7 @@ const Users = ({ users: allUsers, ...rest }) => {
 };
 
 Users.propTypes = {
-    users: PropTypes.array
+    users: PropTypes.oneOfType([PropTypes.object, PropTypes.array])
 };
 
 export default Users;
